Create ThreeInputLEDControl.c
The Assignment: 

Write a program that uses a SPDT switch, a photoresistor, and a button.

The switch will change between the photoresistor and a button.
	Setup: 6 lights in a row: red, blue, green, yellow, red, blue
	
	When the switch is flipped, the green light will light up.  

	When the button is pressed, the green light will turn off and the yellow light will turn on 
for 1 seconds.  Then the yellow light will turn off and the red light will turn on for 3 
seconds.  Then the green light will turn on until the button is pressed again.
Setup:	on:                green				off:red, blue,            yellow, red, blue	
1st: 	on:  	                     yellow			off:red, blue, green              red, blue
2nd: 	on:			          red		off:red, blue, green, yellow,         blue                               
3rd: 	on:                green				off:red, blue,            yellow, red, blue


The photoresistor will begin blinking lights when it is dark, and turn off all lights when it is light.
	
Setup:   6 lights in a row: red, blue, green, yellow, red, blue

Turn on the first red and blue together, then turn on green and yellow together, then turn 
on the last red and blue together.  Turn off the first red, then blue, green, yellow, red, blue.  Repeat.
Setup	on: 						off:red, blue, green, yellow, red, blue
1st: 	on: red, blue					off:                green, yellow, red, blue
2nd: 	on: red, blue, green, yellow			off:                                       red, blue
3rd: 	on: red, blue, green, yellow, red, blue	off: 
4th: 	on:        blue, green, yellow, red, blue		off: red
5th:	on:                 green, yellow, red, blue	off: red, blue
6th: 	on:         	          yellow, red, blue	off: red, blue, green
7th: 	on:			          red, blue	off: red, blue, green, yellow
8th: 	on:				      blue	off: red, blue, green, yellow, red
9th:	on: 						off: red, blue, green, yellow, red, blue



Use a loop to turn the lights on and off.  

Use a loop to setup your lights as outputs.